# backend/api-files/openapi.yaml
openapi: 3.0.3
info:
  title: Hopp API
  description: API specification for Hopp backend services
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  schemas:
...
    PrivateUser:
      allOf:
        - $ref: "#/components/schemas/BaseUser"
        - type: object
          properties:
            social_metadata:
              type: object
              additionalProperties: true
              nullable: true
            metadata:
              type: object
              additionalProperties: true
              nullable: true
            is_pro:
              type: boolean
              description: Whether the user has an active paid subscription
            is_trial:
              type: boolean
              description: Whether the user is currently in trial period
            trial_ends_at:
              type: string
              format: date-time
              nullable: true
              description: When the trial period ends (null if not in trial or has active subscription)

      type: object
...
    Room:
      required:
        - id
        - name
        - user_id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        user_id:
          type: string
...
    Error:
      type: object
      properties:
        message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/auth/social/:provider:
    get:
      summary: Initiate social login with specified provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, slack]
      responses:
        "302":
          description: Redirect to provider's login page
  /api/auth/social/:provider/callback:
    get:
      summary: Social login callback endpoint
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, slack]
      responses:
        "200":
          description: Successful authentication
        "401":
          description: Authentication failed
  /api/sign-up:
    post:
      summary: Manual sign up endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                  description: User's first name
                last_name:
                  type: string
                  description: User's last name
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  minLength: 12
                  maxLength: 72
                  description: User's password (12-72 characters)
                team_name:
                  type: string
                  description: Name of the team (required unless team_invite_uuid is provided)
                team_invite_uuid:
                  type: string
                  format: uuid
                  description: UUID for team invitation (if joining an existing team)
                turnstile_token:
                  type: string
                  description: Cloudflare Turnstile token. Required when Turnstile is enabled server-side; ignored otherwise.
      responses:
        "200":
          description: Successfully signed up
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT authentication token
                  user:
                    $ref: "#/components/schemas/PrivateUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/sign-in:
    post:
      summary: Manual sign in endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
                turnstile_token:
                  type: string
                  description: Cloudflare Turnstile token. Required when Turnstile is enabled server-side; ignored otherwise.
      responses:
        "200":
          description: Successfully signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT authentication token
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many sign-in attempts; retry after the window elapses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/forgot-password:
    post:
      summary: Request password reset email
      description: Sends a password reset link to the specified email address if it exists in the system. Always returns success to avoid user enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                turnstile_token:
                  type: string
                  description: Cloudflare Turnstile token. Required when Turnstile is enabled server-side; ignored otherwise.
      responses:
        "200":
          description: Success message (always returned regardless of whether email exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If the email you specified exists in our system, we've sent a password reset link to it."
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/reset-password/{token}:
    patch:
      summary: Reset password using token
      description: Resets the user's password using a valid password reset token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Password reset token received via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
                  minLength: 12
                  maxLength: 72
                  description: New password (12-72 characters; 72 is bcrypt's byte ceiling)
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Your password has been changed. You can now use it to log in."
        "400":
          description: Invalid input, missing token, invalid/expired token, or token already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/authenticate-app:
    get:
      summary: Create JWT token for desktop app
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/user:
    get:
      summary: Get current user details
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/teammates:
    get:
      summary: Get current user's teammates
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of teammates retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BaseUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/websocket:
    get:
      summary: WebSocket connection endpoint
      security:
        - BearerAuth: []
      responses:
        "101":
          description: Switching protocols to WebSocket
        "401":
          description: Unauthorized
  /api/auth/update-user-name:
    put:
      summary: Update user's first and last name
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
              properties:
                first_name:
                  type: string
                  description: User's first name
                last_name:
                  type: string
                  description: User's last name
      responses:
        "200":
          description: Name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/team:
    patch:
      summary: Update the current admin's team
      description: >-
        Updates the authenticated team admin's team. Currently only the team name
        can be changed. Used by onboarding to rename the placeholder team created
        at signup.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: New team name
      responses:
        "200":
          description: Team updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Invalid input or user is not part of any team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - user is not a team admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/get-invite-uuid:
    get:
      summary: Get or create a team invitation UUID
      description: Returns a UUID that can be used to invite users to the team. If the existing invitation has expired, a new one is created.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Team invitation UUID retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite_uuid:
                    type: string
                    format: uuid
                    description: UUID for team invitation
                  team_name:
                    type: string
                    description: Name of the team
        "400":
          description: User is not part of any team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/send-team-invites:
    post:
      summary: Send team invitation emails
      description: Sends invitation emails to a list of email addresses to join the user's team
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - invitees
              properties:
                invitees:
                  type: array
                  items:
                    type: string
                    format: email
                  description: List of email addresses to invite
      responses:
        "200":
          description: Invitations sent successfully
        "400":
          description: Invalid request or user is not part of any team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/metadata/onboarding-form:
    post:
      summary: Update onboarding form status
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                onboarding:
                  type: object
                  additionalProperties: true
                  description: Free-form onboarding survey answers
      responses:
        "200":
          description: Onboarding form status updated successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/invitation-details/{uuid}:
    get:
      summary: Get team details for an invitation
      description: Returns information about the team associated with an invitation UUID
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Name of the team the user is invited to join
                  team_id:
                    type: integer
                    format: uint
                    description: ID of the team the user is invited to join
        "400":
          description: Invalid invitation UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invitation not found or has expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/room:
    post:
      summary: Create a new room
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the room
      responses:
        "200":
          description: LiveKit tokens retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
                required:
                  - id
                  - name
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/rooms:
    get:
      summary: Get all rooms for the user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Rooms for user retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
...