# backend/internal/server/server.go
package server

import (
	"context"
	"encoding/gob"
	"errors"
	"fmt"
	"hopp-backend/internal/callstate"
	"hopp-backend/internal/common"
	"hopp-backend/internal/config"
	"hopp-backend/internal/email"
	"hopp-backend/internal/handlers"
	"hopp-backend/internal/models"
	"html/template"
	"io"
	"math"
	"net/http"
	"os"
	"strconv"
	"strings"
	"time"

	"github.com/go-playground/validator"
	"github.com/labstack/echo-contrib/echoprometheus"
	"github.com/labstack/echo-contrib/session"
	"github.com/labstack/echo/v4"
	"github.com/labstack/echo/v4/middleware"
	"github.com/labstack/gommon/log"
	"github.com/markbates/goth"
	"github.com/markbates/goth/gothic"
	"github.com/markbates/goth/providers/github"
	"github.com/markbates/goth/providers/google"
	"github.com/markbates/goth/providers/slack"
	"github.com/prometheus/client_golang/prometheus"
	"github.com/redis/go-redis/v9"
	resend "github.com/resend/resend-go/v2"
	"github.com/wader/gormstore/v2"
	"gorm.io/driver/postgres"
	"gorm.io/driver/sqlite"
	"gorm.io/gorm"
	"gorm.io/gorm/logger"
)

// CustomValidator Source: https://echo.labstack.com/docs/request#validate-data
type CustomValidator struct {
	validator *validator.Validate
}

func (cv *CustomValidator) Validate(i interface{}) error {
	if err := cv.validator.Struct(i); err != nil {
		// Optionally, you could return the error to give each route more control over the status code
		return err
	}
	return nil
}

type Template struct {
	templates *template.Template
}

func (t *Template) Render(w io.Writer, name string, data interface{}, c echo.Context) error {
	return t.templates.ExecuteTemplate(w, name, data)
}

type SentryLogger struct {
	echo.Logger
}

func (l *SentryLogger) Error(i ...interface{}) {
	// Capture in Sentry
	if err, ok := i[0].(error); ok {
		handlers.CaptureError(err)
	} else {
		handlers.CaptureError(fmt.Errorf("%v", i...))
	}
	// Call original logger
	l.Logger.Error(i...)
}

type Server struct {
	common.ServerState
}

func New(cfg *config.Config) *Server {
	e := echo.New()
	e.Validator = &CustomValidator{validator: validator.New()}
	e.Logger = &SentryLogger{Logger: e.Logger}
	e.Logger.SetLevel(log.DEBUG)

	return &Server{
		common.ServerState{
			Echo:   e,
			Config: cfg,
		},
	}
}

func (s *Server) Initialize() error {
	// Initialize database
	s.setupDatabase()

	s.setupRedis()

	if s.Redis != nil {
		s.CallState = callstate.NewTracker(s.Redis, s.Echo.Logger)
		// Presence is reconciled against LiveKit as a hint (confirmed by clients
		// before being shown).
		// Run one synchronous pass on startup to seed the
		// snapshot, then start the periodic loop.
		if s.Config.Livekit.ServerURL != "" {
			if err := s.CallState.EnableReconciliation(s.DB, s.Config.Livekit.ServerURL, s.Config.Livekit.APIKey, s.Config.Livekit.Secret); err != nil {
				s.Echo.Logger.Warnf("Failed to enable presence reconciliation: %v", err)
			} else {
				ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
				s.CallState.ReconcileOnce(ctx)
				cancel()
				s.CallState.StartReconciliation()
			}
		} else {
			s.Echo.Logger.Warn("LIVEKIT_SERVER_URL not configured, presence reconciliation disabled")
		}
	}

	// Initialize JWT
	s.JwtIssuer = handlers.NewJwtAuth(s.Config.Auth.SessionSecret)

	// Initialize Resend email client
	s.setupEmailClient()

	// Initialize session store
	s.setupSessionStore()

	// Setup templates
	s.setupTemplates()

	// Setup routes
	s.setupRoutes()

	// Run Migrations
	s.runMigrations()

	// Setup goth providers
	s.setupGothProviders()

	s.setupMetrics()

	// Setup middleware -
	// Keep last to avoid Recover middleware and panic if something goes wrong on init
	s.setupMiddleware()

	return nil
}

func (s *Server) setupDatabase() {
	dsn := s.Config.Database.DSN
	if dsn == "" {
		s.Echo.Logger.Fatal("DATABASE_DSN environment variable is required")
	}

	var db *gorm.DB
	var err error

	// Detect database driver from DSN
	// SQLite DSNs typically start with "file:"
	if strings.HasPrefix(dsn, "file:") {
		// Use SQLite driver for testing
		db, err = gorm.Open(sqlite.Open(dsn), &gorm.Config{TranslateError: true, Logger: logger.Default.LogMode(logger.Silent)})
	} else {
		// Use PostgreSQL driver for production
		db, err = gorm.Open(postgres.Open(dsn), &gorm.Config{TranslateError: true, Logger: logger.Default.LogMode(logger.Silent)})
	}

	if err != nil {
		s.Echo.Logger.Fatal(err)
	}
	s.DB = db
}

func (s *Server) setupRedis() {
	url := s.Config.Database.RedisURI

	// Make Redis optional - if URI is empty, skip Redis setup
	if url == "" {
		s.Echo.Logger.Warn("REDIS_URI not configured, Redis features will be disabled")
		s.Redis = nil
		return
	}

	opts, err := redis.ParseURL(url)
	if err != nil {
		s.Echo.Logger.Warnf("Failed to parse Redis URL: %v, Redis features will be disabled", err)
		s.Redis = nil
		return
	}

	s.Redis = redis.NewClient(opts)

	// Validate proper connection, but don't panic on failure
	ctx := context.Background()
	result := s.Redis.Ping(ctx)
	if result.Err() != nil {
		s.Echo.Logger.Warnf("Redis connection failed: %v, Redis features will be disabled", result.Err())
		s.Redis = nil
		return
	}
}

func (s *Server) setupSessionStore() {
	store := gormstore.New(s.DB, []byte(s.Config.Auth.SessionSecret))
	store.SessionOpts.MaxAge = 60 * 60 * 24 * 30 // 30 days
	store.SessionOpts.SameSite = http.SameSiteLaxMode
	store.SessionOpts.HttpOnly = true

	quit := make(chan struct{})
	go store.PeriodicCleanup(1*time.Hour, quit)

	// To solve securecookie: error - caused by: gob: type not registered for interface
	gob.Register(map[string]interface{}{})

	s.Store = store
}

func (s *Server) setupTemplates() {
	// Try to load templates, but don't fail if they don't exist (e.g., in tests)
	tmpl, err := template.ParseGlob("./web/*.html")
	if err != nil {
		s.Echo.Logger.Warnf("Failed to load templates: %v, template rendering will be disabled", err)
		return
	}
	t := &Template{
		templates: tmpl,
	}
	s.Echo.Renderer = t
}

func (s *Server) runMigrations() {
	err := s.DB.AutoMigrate(
		&models.User{},
		&models.Team{},
		&models.Room{},
		&models.ResetToken{},
		&models.TeamInvitation{},
		&models.EmailInvitation{},
		&models.Subscription{},
		&models.Feedback{},
		&models.SlackInstallation{},
	)
	if err != nil {
		s.Echo.Logger.Fatal(err)
	}
}

func (s *Server) setupMiddleware() {
	s.Echo.Use(middleware.CORS())
	s.Echo.Use(session.Middleware(s.Store))
	s.Echo.Use(middleware.Recover())
	// Try to add prometheus middleware, but don't panic if already registered (e.g., in tests)
	// This allows multiple test runs without panicking
	defer func() {
		if r := recover(); r != nil {
			if err, ok := r.(error); ok && err.Error() == "duplicate metrics collector registration attempted" {
				s.Echo.Logger.Warn("Prometheus middleware already registered, skipping")
			} else {
				panic(r)
			}
		}
	}()
	s.Echo.Use(echoprometheus.NewMiddleware("renkey_backend"))
}

func (s *Server) setupMetrics() {
	// Only register Redis metrics if Redis is available
	if s.Redis == nil {
		return
	}

	// Collectors register into Prometheus's global registry, so initializing more
	// than one Redis-enabled server in a single process (e.g. across tests) would
	// otherwise fail on a duplicate registration. Register (unlike MustRegister)
	// returns the error, so we can tolerate the already-registered case.
	register := func(c prometheus.Collector) {
		err := prometheus.Register(c)
		if err == nil {
			return
		}
		var alreadyRegistered prometheus.AlreadyRegisteredError
		if errors.As(err, &alreadyRegistered) {
			s.Echo.Logger.Warn("Prometheus collector already registered, skipping")
			return
		}
		s.Echo.Logger.Errorf("Failed to register Prometheus collector: %v", err)
	}

	register(prometheus.NewGaugeFunc(
		prometheus.GaugeOpts{
			Subsystem: "redis",
			Name:      "connected_clients",
			Help:      "The number of clients currently connected to Redis",
		},
		func() float64 {
			ctx := context.Background()
			connectedClientsRaw := s.Redis.InfoMap(ctx).Item("Clients", "connected_clients")

			connectedClients, err := strconv.ParseFloat(connectedClientsRaw, 64)
			if err != nil {
				return math.NaN()
			}

			return connectedClients
		},
	))

	if s.CallState != nil {
		register(prometheus.NewGaugeFunc(
			prometheus.GaugeOpts{
				Subsystem: "callstate",
				Name:      "active_rooms",
				Help:      "Number of active call rooms in the presence snapshot",
			},
			func() float64 {
				ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
				defer cancel()
				return float64(s.CallState.RoomCount(ctx))
			},
		))
	}
}

func (s *Server) setupGothProviders() {
	// Set the session secret for Goth
	gothic.Store = s.Store

	goth.UseProviders(
		google.New(s.Config.Auth.GoogleKey, s.Config.Auth.GoogleSecret, s.Config.Auth.GoogleRedirect, "email", "profile", "openid"),
		slack.New(s.Config.Auth.SlackKey, s.Config.Auth.SlackSecret, s.Config.Auth.SlackRedirect, "users:read", "users:read.email", "team:read"),
		github.New(s.Config.Auth.GitHubKey, s.Config.Auth.GitHubSecret, s.Config.Auth.GitHubRedirect, "user:email", "read:user"),
	)
}

func (s *Server) setupEmailClient() {
	apiKey := s.Config.Resend.APIKey
	if apiKey == "" {
		s.Echo.Logger.Warn("RESEND_API_KEY not configured, email notifications will be disabled")
		return
	}

	resendClient := resend.NewClient(apiKey)
	s.EmailClient = email.NewResendEmailClient(resendClient,
		s.Config.Resend.DefaultSender,
		s.Echo.Logger)
}

func (s *Server) setupRoutes() {
	handlers.SetupSentry(s.Echo, s.Config)

	// Serve static files
	s.Echo.Static("/static", "web/static")

	// Initialize handlers
	auth := handlers.NewAuthHandler(s.DB, s.Config, s.JwtIssuer, s.Redis, &handlers.RealGothicProvider{})
	billing := handlers.NewBillingHandler(s.DB, s.Config, s.JwtIssuer, s.EmailClient)
	slackHndlr := handlers.NewSlackHandler(s.DB, s.Config, s.JwtIssuer, s.Redis, s.Echo.Logger)

	// Set the EmailClient field directly
	auth.EmailClient = s.EmailClient
	auth.CallState = s.CallState
	slackHndlr.CallState = s.CallState

	// Start the background Slack room cleanup goroutine
	// This runs every 5 seconds and cleans up empty Slack rooms
	slackHndlr.StartSlackRoomCleanup()

	// API routes group
	api := s.Echo.Group("/api")

	// Public API endpoints
	api.GET("/health", func(c echo.Context) error {
		return c.String(200, "OK")
	})
	api.GET("/metrics", echoprometheus.NewHandler())
	// Add invitation details endpoint
	api.GET("/invitation-details/:uuid", auth.GetInvitationDetails)

	// Unsubscribe endpoints
	api.GET("/unsubscribe/:token", auth.UnsubscribeUser)
	api.POST("/unsubscribe/:token", auth.UnsubscribeUser)

	// Billing webhook endpoint (public)
	api.POST("/billing/webhook", billing.HandleWebhook)

	// Slack app endpoints (public - verified via Slack signature)
	api.GET("/slack/oauth/callback", slackHndlr.SlackOAuthCallback)
	api.POST("/slack/commands/hopp", slackHndlr.HandleHoppCommand)
	api.POST("/slack/interactions", slackHndlr.HandleInteraction)
	// Slack join session endpoint (public, validates room and redirects to web app)
	api.GET("/slack/join/:sessionId", slackHndlr.JoinPairingSession)

	// Authentication endpoints
	api.GET("/auth/social/:provider", auth.SocialLogin)
	api.GET("/auth/social/:provider/callback", auth.SocialLoginCallback)
	api.POST("/sign-up", auth.ManualSignUp)
	api.POST("/sign-in", auth.ManualSignIn)
	api.POST("/forgot-password", auth.ForgotPassword)
	api.PATCH("/reset-password/:token", auth.ResetPassword)

	// Protected API routes group
	protectedAPI := api.Group("/auth", s.JwtIssuer.Middleware())

	protectedAPI.GET("/authenticate-app", auth.AuthenticateApp)
	protectedAPI.GET("/user", auth.User)
	protectedAPI.PUT("/update-user-name", auth.UpdateName)
	protectedAPI.PATCH("/team", auth.UpdateTeam)
	protectedAPI.GET("/teammates", auth.Teammates)
	protectedAPI.DELETE("/teammates/:userId", auth.RemoveTeammate)

	protectedAPI.GET("/websocket", handlers.CreateWSHandler(&s.ServerState))
	protectedAPI.GET("/get-invite-uuid", auth.GetInviteUUID)
	protectedAPI.POST("/change-team/:uuid", auth.ChangeTeam)
	protectedAPI.POST("/send-team-invites", auth.SendTeamInvites)
	protectedAPI.POST("/metadata/onboarding-form", auth.UpdateOnboardingFormStatus)
	protectedAPI.POST("/subscribe-linux-waitlist", auth.SubscribeToLinuxWaitingList)

	protectedAPI.POST("/room", auth.CreateRoom)
	protectedAPI.PUT("/room/:id", auth.UpdateRoom)
	protectedAPI.DELETE("/room/:id", auth.DeleteRoom)
	protectedAPI.GET("/room/:id", auth.GetRoom)
	protectedAPI.GET("/rooms", auth.GetRooms)
	protectedAPI.GET("/rooms/presence", auth.GetRoomsPresence)
	protectedAPI.GET("/calls/presence", auth.GetCallsPresence)
	protectedAPI.POST("/call/join/:userId", auth.JoinCall)
	protectedAPI.POST("/room/:id/leave", slackHndlr.LeaveRoom)

	// LiveKit server endpoint
	protectedAPI.GET("/livekit/server-url", auth.GetLivekitServerURL)

	// Slack session tokens endpoint (protected - requires JWT)
	// This is called by the desktop app to get tokens for joining a Slack session
	protectedAPI.GET("/slack/session/:sessionId/tokens", slackHndlr.GetSessionTokens)

	// Slack installation management endpoints (protected - requires JWT)
	protectedAPI.GET("/slack/install", slackHndlr.SlackInstall)
	protectedAPI.GET("/slack/installation", slackHndlr.GetSlackInstallation)
	protectedAPI.DELETE("/slack/installation", slackHndlr.DeleteSlackInstallation)

	// Feedback endpoint
	protectedAPI.POST("/feedback", auth.SubmitFeedback)

	// Protected billing endpoints
	protectedAPI.GET("/billing/subscription", billing.GetSubscriptionStatus)
	protectedAPI.POST("/billing/create-checkout-session", billing.CreateCheckoutSession)
	protectedAPI.POST("/billing/create-portal-session", billing.CreatePortalSession)
	protectedAPI.GET("/billing/settings", billing.GetBillingSettings)
	protectedAPI.PUT("/billing/settings", billing.UpdateBillingSettings)

	// Debug endpoints - only enabled when ENABLE_DEBUG_ENDPOINTS=true
	if s.Config.Server.Debug {
		api.GET("/debug", func(c echo.Context) error {
			return c.Render(http.StatusOK, "debug.html", nil)
		})
		api.GET("/call-token", auth.GenerateDebugCallToken)
		api.GET("/jwt-debug", func(c echo.Context) error {
			email := c.QueryParam("email")
			token, err := s.JwtIssuer.GenerateToken(email)
			if err != nil {
				return c.String(http.StatusInternalServerError, "Failed to generate token")
			}
			return c.JSON(http.StatusOK, map[string]string{
				"email": email,
				"token": token,
			})
		})
	}

	// SPA handler - serve index.html for all other routes
	s.Echo.GET("/*", func(c echo.Context) error {
		// Skip API routes
		if strings.HasPrefix(c.Request().URL.Path, "/api") {
			return echo.NewHTTPError(http.StatusNotFound, "API endpoint not found")
		}
		webAppPath := "web/web-app.html"
		if s.Config.Server.Debug {
			webAppPath = "web/web-app-debug.html"
		}
		return c.File(webAppPath)
	})
}

func (s *Server) Start() error {
	serverURL := s.Config.Server.Host + ":" + s.Config.Server.Port

	if s.Config.Server.TLS.Enabled {
		if _, err := os.Stat(s.Config.Server.TLS.CertFile); os.IsNotExist(err) {
			s.Echo.Logger.Warn("TLS certificate file not found, falling back to HTTP")
			return s.Echo.Start(serverURL)
		}
		if _, err := os.Stat(s.Config.Server.TLS.KeyFile); os.IsNotExist(err) {
			s.Echo.Logger.Warn("TLS key file not found, falling back to HTTP")
			return s.Echo.Start(serverURL)
		}
		return s.Echo.StartTLS(serverURL, s.Config.Server.TLS.CertFile, s.Config.Server.TLS.KeyFile)
	}

	return s.Echo.Start(serverURL)
}